refactor: file structure of pages and components folders
