+++ src/components/MakeWordGame/WordInfo.tsx
